Rhubarb Accounts/SpendTracker-AzureFunction/requirements.txt
azure-functions==1.18.0
azure-functions-worker==1.0.15
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pyodbc==5.0.1
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
python-dateutil==2.8.2
requests==2.31.0